+++ adminalerts/tests/test_ui.py
@@ -33,32 +33,55 @@ class AdminAlertUITestBase(AdminAlertMixin, UITestBase):
 class TestAlertMessage(AdminAlertUITestBase):
     """Tests for the admin alert message"""
 
+    def setUp(self):
+        super().setUp()
+        self.url = reverse('home')
+
     def test_message(self):
         """Test visibility of alert message in home view"""
         expected = [(self.superuser, 1), (self.regular_user, 1)]
-        url = reverse('home')
         self.assert_element_count(
-            expected, url, 'sodar-alert-site-app', 'class'
+            expected, self.url, 'sodar-alert-site-app', 'class'
+        )
+        self.assert_element_count(
+            expected, self.url, 'sodar-alert-full-text-link', 'class'
+        )
+        self.assert_element_count(
+            expected, self.url, 'sodar-alert-detail-link', 'class'
+        )
+
+    def test_message_no_description(self):
+        """Test alert message without description"""
+        self.alert.description = ''
+        self.alert.save()
+        expected = [(self.superuser, 1), (self.regular_user, 1)]
+        self.assert_element_count(
+            expected, self.url, 'sodar-alert-site-app', 'class'
+        )
+        expected = [(self.superuser, 0), (self.regular_user, 0)]
+        self.assert_element_count(
+            expected, self.url, 'sodar-alert-full-text-link', 'class'
+        )
+        self.assert_element_count(
+            expected, self.url, 'sodar-alert-detail-link', 'class'
         )
 
     def test_message_inactive(self):
-        """Test visibility of an inactive alert message"""
+        """Test visibility of inactive alert message"""
         self.alert.active = 0
         self.alert.save()
         expected = [(self.superuser, 0), (self.regular_user, 0)]
-        url = reverse('home')
         self.assert_element_count(
-            expected, url, 'sodar-alert-site-app', 'class'
+            expected, self.url, 'sodar-alert-site-app', 'class'
         )
 
     def test_message_expired(self):
-        """Test visibility of an expired alert message"""
+        """Test visibility of expired alert message"""
         self.alert.date_expire = timezone.now() - timezone.timedelta(days=1)
         self.alert.save()
         expected = [(self.superuser, 0), (self.regular_user, 0)]
-        url = reverse('home')
         self.assert_element_count(
-            expected, url, 'sodar-alert-site-app', 'class'
+            expected, self.url, 'sodar-alert-site-app', 'class'
         )
 
     def test_message_login(self):
@@ -71,26 +94,29 @@ class TestAlertMessage(AdminAlertUITestBase):
         """Test visibility of alert in login view without auth requirement"""
         self.alert.require_auth = False
         self.alert.save()
-
         self.selenium.get(self.build_selenium_url(reverse('login')))
         self.assertIsNotNone(
             self.selenium.find_element(By.CLASS_NAME, 'sodar-alert-site-app')
         )
 
 
-class TestListView(AdminAlertUITestBase):
-    """Tests for the admin alert list view"""
+class TestAdminAlertListView(AdminAlertUITestBase):
+    """Tests for AdminAlertListView"""
+
+    def setUp(self):
+        super().setUp()
+        self.url = reverse('adminalerts:list')
 
     def test_list_items(self):
         """Test existence of items in list"""
         expected = [(self.superuser, 1)]
-        url = reverse('adminalerts:list')
-        self.assert_element_count(expected, url, 'sodar-aa-alert-item', 'id')
+        self.assert_element_count(
+            expected, self.url, 'sodar-aa-alert-item', 'id'
+        )
 
     def test_list_dropdown(self):
         """Test existence of alert dropdown in list"""
         expected = [(self.superuser, 1)]
-        url = reverse('adminalerts:list')
         self.assert_element_count(
-            expected, url, 'sodar-aa-alert-dropdown', 'class'
+            expected, self.url, 'sodar-aa-alert-dropdown', 'class'
         )